+++ core/src/net/hollowbit/archipelo/network/packets/PlayerStatsPacket.java
@@ -5,7 +5,7 @@ import net.hollowbit.archipelo.network.PacketType;
 
 public class PlayerStatsPacket extends Packet {
 
-	public int health = 0;
+	public float health = 0;
 	
 	public PlayerStatsPacket() {
 		super(PacketType.PLAYER_STATS);
